+++ src/main/java/com/redhat/cloud/notifications/auth/rhid/RhIdPrincipal.java
@@ -1,4 +1,4 @@
-package com.redhat.cloud.notifications.auth;
+package com.redhat.cloud.notifications.auth.rhid;
 
 import java.security.Principal;
