+++ src/main/java/com/redhat/cloud/notifications/auth/rhid/RhIdentityAuthenticationRequest.java
@@ -1,8 +1,8 @@
-package com.redhat.cloud.notifications.auth;
+package com.redhat.cloud.notifications.auth.rhid;
 
 import io.quarkus.security.identity.request.BaseAuthenticationRequest;
 
-import static com.redhat.cloud.notifications.auth.RHIdentityAuthMechanism.IDENTITY_HEADER;
+import static com.redhat.cloud.notifications.auth.rhid.RHIdentityAuthMechanism.IDENTITY_HEADER;
 
 public class RhIdentityAuthenticationRequest extends BaseAuthenticationRequest {
